Set container width based on dropdown contents
The width was being set to the same as the original select field, which
can vary wildly based on the OS and browser, and meant that the Chosen
field couldn't look all that different from the original. Increasing the
font size or padding by much would cause cropping.

Instead, this just sizes the container by measuring the width of the
dropdown after it's been populated the first time.

I haven't tested this widely yet, and I'm aware that the search box is
too big. Also, I'm aware that this is not the upstream chosen repo. I'm
just hacking this thing together.